Getting started on the engine

+++ include/robin/generation/transformation.h
@@ -8,8 +8,8 @@
 namespace ff {
   struct Transformation {
     Affine factors_{};
-    double weight_{};  // probabiliy that this transformation is picked
-    double color_{}; // 0.0 to 1.0 representing the index of a gradient
+    float weight_{};  // proportion of time that this transformation is picked
+    float color_{}; // 0.0 to 1.0 representing the index of a gradient
     // each transformation is the weighted sum of multiple variations
     std::vector<Variation> variations_{};
   };
